Simplifying the modules.
