+++ app/src/androidTest/java/com/mycompany/myapp/MainApplicationDaggerMockRule.java
@@ -1,9 +1,10 @@
 package com.mycompany.myapp;
 
 import com.mycompany.myapp.app.AndroidModule;
+import com.mycompany.myapp.app.AppModule;
 import com.mycompany.myapp.app.ApplicationComponent;
+import com.mycompany.myapp.app.VariantModule;
 import com.mycompany.myapp.data.DataModule;
-import com.mycompany.myapp.data.api.github.GitHubModule;
 import com.mycompany.myapp.modules.CrashReporterModule;
 import com.mycompany.myapp.monitoring.LoggerModule;
 
@@ -17,11 +18,12 @@ public class MainApplicationDaggerMockRule extends DaggerMockRule<ApplicationCom
         // included in this list of modules.
         // In other works, it is best to list out all of your app modules.
         super(ApplicationComponent.class,
+                new VariantModule(),
                 new AndroidModule(getAppUnderTest()),
+                new AppModule(),
                 new LoggerModule(),
                 new CrashReporterModule(),
-                new DataModule(),
-                new GitHubModule());
+                new DataModule());
         set(component -> getAppUnderTest().setComponent(component));
     }
 }
