+++ app/src/main/java/com/mycompany/myapp/data/DataModule.java
@@ -2,6 +2,12 @@ package com.mycompany.myapp.data;
 
 import android.app.Application;
 
+import com.google.gson.Gson;
+import com.mycompany.myapp.app.Settings;
+import com.mycompany.myapp.data.api.Api;
+import com.mycompany.myapp.data.api.github.GitHubApiService;
+import com.mycompany.myapp.data.api.github.GitHubService;
+
 import java.io.File;
 
 import javax.inject.Singleton;
@@ -10,6 +16,10 @@ import dagger.Module;
 import dagger.Provides;
 import okhttp3.Cache;
 import okhttp3.OkHttpClient;
+import retrofit2.Converter;
+import retrofit2.Retrofit;
+import retrofit2.adapter.rxjava.RxJavaCallAdapterFactory;
+import retrofit2.converter.gson.GsonConverterFactory;
 
 @Module
 public class DataModule {
@@ -28,5 +38,47 @@ public class DataModule {
         return new OkHttpClient.Builder()
                 .cache(cache)
                 .build();
+    }
+
+    @Singleton
+    @Api("github")
+    @Provides
+    public String provideBaseUrl(Settings settings) {
+        return settings.getBaseUrl();
+    }
+
+    @Singleton
+    @Api("github")
+    @Provides
+    public Converter.Factory provideConverter() {
+        Gson gson = new Gson();
+        return GsonConverterFactory.create(gson);
+    }
+
+    @Singleton
+    @Api("github")
+    @Provides
+    public Retrofit provideRetrofit(
+            OkHttpClient client,
+            @Api("github") String baseUrl,
+            @Api("github") Converter.Factory converterFactory) {
+        return new Retrofit.Builder()
+                .client(client)
+                .baseUrl(baseUrl)
+                .addConverterFactory(converterFactory)
+                .addCallAdapterFactory(RxJavaCallAdapterFactory.create())
+                .build();
+    }
+
+    @Singleton
+    @Provides
+    public GitHubApiService provideGitHubApiService(@Api("github") Retrofit retrofit) {
+        return retrofit.create(GitHubApiService.class);
+    }
+
+    @Singleton
+    @Provides
+    public GitHubService provideGitHubService(GitHubApiService api) {
+        return new GitHubService(api);
     }
 }
